moved related tests to different files
